requirements.txt
instapy==0.6.13
selenium==3.141.0
python-dotenv==0.19.0
webdriver-manager==3.4.2
requests==2.26.0